% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCoverage.R
\name{checkFreqCovPCC}
\alias{checkFreqCovPCC}
\title{Quality control of coverage cormalization result}
\usage{
checkFreqCovPCC(df, outpdfname = "NULL")
}
\description{
Takes coverage normalized result and check PCC between ligation frequency and coverage values
}
