% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcDistance.R
\name{checkFreqDistPCC}
\alias{checkFreqDistPCC}
\title{Quality control of distance cormalization result}
\usage{
checkFreqDistPCC(df, outpdfname = "NULL")
}
\description{
Takes distance normalized result and check PCC between ligation frequency and distance values
}
