% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactPval.R
\name{contactPval}
\alias{contactPval}
\title{Compute p-value/FDR of normalized interaction by 3P-Weibull fitting}
\usage{
contactPval(df, outpdfname)
}
\arguments{
\item{df}{A coverage-distance normalized output df.}

\item{outpdfname}{fitDistr-generated fitting result pdf name.}
}
\value{
df Added p-value/FDR of interactions as columns.
}
\description{
Takes formatted dataframe and retunrs dataframe with 3P Weibull fitted value
}
