% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInputDF.R
\name{filterInputDF}
\alias{filterInputDF}
\title{Filter input data frame rows}
\usage{
filterInputDF(
  df,
  check_trans_values = TRUE,
  remove_zero_contact = TRUE,
  self_ligation_distance = 15000
)
}
\arguments{
\item{df}{Input data frame.}

\item{check_trans_values}{If TRUE, rows containing trans interactions dropped.}

\item{remove_zero_contact}{If TRUE, rows wifh $freq value 0 dropped.}

\item{self_ligation_distance}{Distance between two fragments regarded as self-ligation. Default 15kb.}
}
\value{
df with rows filtered
}
\description{
Takes formatted dataframe and filtered df
}
