% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCoverage.R
\name{normCoverage}
\alias{normCoverage}
\title{Coverage Normalization function}
\usage{
normCoverage(df, do_shuffle = TRUE, cov1_thresh = 200, cov2_thresh = 200,
  max_covnorm_value = 50, sample_ratio = -1)
}
\arguments{
\item{df}{An input dataframe.}

\item{max_covnorm_value}{Outliers by division with too small a number are replaced by this value.}

\item{sample_ratio}{If NOT -1 && between 0.0-1.0, downsample the df rows to be used for fitting by that ratio.}

\item{cov1_thresh&cov2_thresh}{Threshold for filteraing bins with too low coverage.}
}
\value{
df Coverage normalized.
}
\description{
Takes formatted dataframe and retunrs glm.nb fitted coeffs & normalized df
}
