% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDistance.R
\name{normDistance}
\alias{normDistance}
\title{Distance normalization function}
\usage{
normDistance(df, max_dist = -1, max_distnorm_value = 50, sample_ratio = -1)
}
\arguments{
\item{df}{An input dataframe, coverage normalized.}

\item{max_dist}{If > 0, Rows with larger $dist values will be dropped.}

\item{max_distnorm_value}{Outliers by division with too small a number are replaced by this value.}

\item{sample_ratio}{If NOT -1 && between 0.0-1.0, downsample the df rows to be used for fitting by that ratio.}
}
\value{
df Distance normalized.
}
\description{
Takes formatted dataframe and retunrs glm.nb fitted coeffs & normalized df
}
