% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCoverage.R
\name{plotCovNormRes}
\alias{plotCovNormRes}
\title{Plot coverage-interaction frequency before/expected/after normalization}
\usage{
plotCovNormRes(df, binsize = 100, pct_quantile = 0.99,
  outpdfname = "plot_coverage_norm_result.pdf")
}
\description{
Takes coverage normalized result and plot 3 heatmaps
}
