% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcDistance.R
\name{plotDistNormRes}
\alias{plotDistNormRes}
\title{Plot distance-interaction frequency before/after normalization}
\usage{
plotDistNormRes(
  df,
  maxX = 2e+06,
  maxY1 = 70,
  maxY2 = 15,
  outpdfname = "plot_distance_norm_result.pdf"
)
}
\description{
Takes distance normalized result and plot two hexagonal heatmap
}
